+++ MetalLink.Application/Interfaces/IBuyerRepository.cs
@@ -8,8 +8,8 @@ namespace MetalLink.Application.Interfaces;
 
 public interface IBuyerRepository
 {
-    Task<Buyer?> GetByIdAsync(long buyerId);
-    Task<Buyer?> GetByIdAsync(long buyerId, CancellationToken cancellationToken);
+    Task<Buyer?> GetByIdAsync(int buyerId);
+    Task<Buyer?> GetByIdAsync(int buyerId, CancellationToken cancellationToken);
     Task<Buyer?> GetByAccountNumberAsync(long accountNumber, CancellationToken cancellationToken = default);
     Task AddAsync(Buyer buyer, CancellationToken cancellationToken = default);
 
@@ -41,8 +41,8 @@ public interface IBuyerRepository
     Task<long> GetNextAccountNumberAsync(CancellationToken ct);
 
     Task<IReadOnlyList<Buyer>> SearchBuyersWithZeroSendingTicketsAsync(
-        long? companyId,
-        long? siteId,
+        int? companyId,
+        int? siteId,
         int? buyerId,
         string? firstName,
         string? lastName,
