+++ MetalLink.Application/Interfaces/IStockMovementRepository.cs
@@ -6,7 +6,7 @@ namespace MetalLink.Application.Interfaces;
 public interface IStockMovementRepository
 {
     Task AddAsync(
-        long productId,
+        int productId,
         decimal baseWeightKg,
         decimal buyWeightKg,
         decimal sellWeightKg,
