+++ MetalLink.Application/Interfaces/ITicketReceivingRepository.cs
@@ -7,18 +7,18 @@ namespace MetalLink.Application.Interfaces;
 
 public interface ITicketReceivingRepository
 {
-    Task<TicketReceiving?> GetByIdAsync(long ticketReceivingId);
+    Task<TicketReceiving?> GetByIdAsync(int ticketReceivingId);
     Task<TicketReceiving?> GetByTicketNumberAsync(string ticketNumber);
     Task<IEnumerable<TicketReceiving>> SearchAsync(
         string? searchTerm = null,
-        long? companyId = null,
-        long? siteId = null,
-        long? customerId = null,
+        int? companyId = null,
+        int? siteId = null,
+        int? customerId = null,
         string? firstName = null,
         string? lastName = null,
         string? idNumber = null,
         long? accountNumber = null,
-        long? productId = null,
+        int? productId = null,
         string? ticketType = null,
         DateTimeOffset? startDate = null,
         DateTimeOffset? endDate = null,
@@ -27,20 +27,20 @@ public interface ITicketReceivingRepository
         int pageSize = 50);
     Task<long> GetCountAsync(
         string? searchTerm = null,
-        long? companyId = null,
-        long? siteId = null,
-        long? customerId = null,
+        int? companyId = null,
+        int? siteId = null,
+        int? customerId = null,
         string? firstName = null,
         string? lastName = null,
         string? idNumber = null,
         long? accountNumber = null,
-        long? productId = null,
+        int? productId = null,
         DateTimeOffset? startDate = null,
         DateTimeOffset? endDate = null,
         string? deliveryStatus = null);
     Task<TicketReceiving> AddAsync(TicketReceiving ticket);
     Task UpdateAsync(TicketReceiving ticket);
-    Task<string> GenerateTicketNumberAsync(long siteId);
+    Task<string> GenerateTicketNumberAsync(int siteId);
     Task<string?> GetLastTicketNumberByPrefixAsync(string prefix);
     Task<long> GetNextTicketSequenceValueAsync(string prefix);
     Task<long> PeekNextTicketSequenceValueAsync(string prefix);
