+++ MetalLink.Shared/Companies/CompanyDto.cs
@@ -1,6 +1,6 @@
 public sealed class CompanyDto
 {
-    public long   CompanyId   { get; set; }
+    public int CompanyId { get; set; }
     public string CompanyName { get; set; } = string.Empty;
     public string? VatNumber  { get; set; }
     public bool IsActive { get; set; } = true;
