+++ MetalLink.Shared/Companies/CompanyLookupDto.cs
@@ -3,7 +3,7 @@ namespace MetalLink.Shared.Companies;
 
 public sealed class CompanyLookupDto
 {
-    public long   CompanyId   { get; set; }
+    public int CompanyId { get; set; }
     public string CompanyName { get; set; } = string.Empty;
 
     public string? VatNumber { get; set; }
