+++ MetalLink.Shared/Tickets/Receiving/TicketReceivingDtos.cs
@@ -106,8 +106,8 @@ public class TicketReceivingSearchRequestDto
     public bool NewCustomerOnly { get; set; }
 
     public string? SearchTerm { get; set; }
-    public long? CompanyId { get; set; }
-    public long? SiteId { get; set; }
+    public int? CompanyId { get; set; }
+    public int? SiteId { get; set; }
     public int? CustomerId { get; set; }
     public string? FirstName { get; set; }
     public string? LastName { get; set; }
